PratikKharad changes

+++ MachineTest/Models/Items.cs
@@ -4,6 +4,6 @@
     {
         public int Quantity { get; set; }
         public string ItemName { get; set; }
-        public float ItemPrice { get; set; }
+        public decimal ItemPrice { get; set; }
     }
 }
